requirements.txt
Pillow
markdown
weasyprint
BeautifulSoup4
segno